move cli date helpers into dates.py
