+++ cli/dates.py
@@ -1,4 +1,6 @@
 from datetime import datetime
+import click
+from stonks import exceptions
 from cli import helpers
 import utils
 
@@ -26,7 +28,7 @@ def handle_start_date(settings):
     else:
         default_date = settings.start_date
 
-    helpers.set_date(
+    set_date(
         settings, default_date,
         'start_date', "Change Start Date", description())
 
@@ -40,6 +42,35 @@ def handle_end_date(settings):
         default_date = settings.end_date
 
     # try:
-    helpers.set_date(
+    set_date(
         settings, default_date, 'end_date', "Change End Date", description())
     return False
+
+
+def set_date(settings, default, field_name, header=None, description=None):
+    is_done = False
+    while not is_done:
+        helpers.pre_menu(settings, header, description)
+        try:
+            datestr = click.prompt(
+                "Enter your date", default=default.strftime("%Y-%m-%d"))
+
+            if field_name == 'start_date':
+                settings.start_date = datestr
+            elif field_name == 'end_date':
+                settings.end_date = datestr
+            else:
+                raise ValueError(
+                    "Wrong Field name provided to cli.cli:set_date")
+            is_done = True
+        except exceptions.DateException:
+            click.echo(utils.cli.format(
+                "\n{Invalid format:red}. Should be one of "
+                f"{get_date_format_str(settings)}\n"))
+            if not click.confirm("Invalid date format, Try again?"):
+                is_done = True
+
+
+def get_date_format_str(settings):
+    return ', '.join(
+        [utils.cli.highlight(f, 'cyan') for f in settings.VALID_DATES_FORMAT])
